v2.1.4 Core
The Fuse "Core" SDK version is a light-weight alternative to the
standard Fuse SDK, and includes a pre-bundled version of AdRally
optimized for serving video ads (both rewarded and non-rewarded). This
SDK is useful for developers who want to keep the final binary size of
their apps as small as possible.